requirements.txt
fastapi
uvicorn
black
python-multipart
pdfplumber
python-dotenv
openai
sqlmodel